% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{barplot.BMORSENV}
\alias{barplot.BMORSENV}
\title{barplot BMORSENV graph}
\usage{
\method{barplot}{BMORSENV}(height, select = "Pearson", ...)
}
\arguments{
\item{height}{\code{BMORSENV object} BMORSENV object, result of use the MTME() function}

\item{select}{\code{character} By default ('Pearson'), plot the Pearson Correlations of the MTME Object, else ('MAAPE'), plot the MAAPE of the BMORSENV Object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot from BMORSENV object
}
