% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{barplot.BMTMERSENV}
\alias{barplot.BMTMERSENV}
\title{barplot BMTMERSENV graph}
\usage{
\method{barplot}{BMTMERSENV}(x, select = "Pearson", ...)
}
\arguments{
\item{x}{\code{BMTMERSENV object} BMTMERSENV object, result of use the MTME() function}

\item{select}{\code{character} By default ('Pearson'), plot the Pearson Correlations of the MTME Object, else ('MSEP'), plot the MSEP of the BMTMERSENV Object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot from BMTMERSENV object
}
