% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BME.R
\name{BME}
\alias{BME}
\title{BME}
\usage{
BME(Y, Z1, nIter, burnIn, thin, bs)
}
\arguments{
\item{Y}{no definition available.}

\item{Z1}{no definition available.}

\item{nIter}{no definition available.}

\item{burnIn}{no definition available.}

\item{thin}{no definition available.}

\item{bs}{no definition available.}
}
\value{

}
\description{
BME
}
