% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMORS.R
\name{BMORS}
\alias{BMORS}
\title{Bayesian Multi-Output Regression Stacking (BMORS)}
\usage{
BMORS(Y = NULL, ETA = NULL, covModel = "BRR",
  predictor_Sec_complete = FALSE, nIter = 2500, burnIn = 500,
  thin = 5, progressBar = TRUE, testingSet = NULL,
  parallelCores = 1, digits = 4)
}
\arguments{
\item{Y}{\code{(matrix)} Phenotypic response where each column is a different trait}

\item{ETA}{\code{(matrix)} This is a two-level list used to specify the regression function (or linear predictor).}

\item{covModel}{\code{(string)} Name of the covariates model to implement (BRR, BayesA, BayesB, BayesC).}

\item{predictor_Sec_complete}{FALSE by default}

\item{nIter}{\code{(integer)} Number of iterations to fit the model.}

\item{burnIn}{\code{(integer)} Number of items to burn at the beginning of the model.}

\item{thin}{\code{(integer)} Number of items to thin the model.}

\item{progressBar}{\code{(Logical)} Show the progress bar.}

\item{testingSet}{\code{(object or vector)} Crossvalidation object or vector with the positions to use like testing in a cross-validation test.}

\item{parallelCores}{\code{(integer)} Number of cores to use.}

\item{digits}{\code{(integer)} Number of digits of accuracy in the results.}
}
\description{
Bayesian Multi-Output Regression Stacking (BMORS)
}
\examples{
\dontrun{
data('wheat')
CV.RandomPart(phenoWheat, NPartitions = 10, PTesting = 0.2, set_seed = 123)
}
}
