% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMORSEnv.R
\name{BMORS_Env}
\alias{BMORS_Env}
\title{Bayesian Multi-Output regression stacking for specific environment estimations}
\usage{
BMORS_Env(data = NULL, testingEnv = "", ETA = NULL,
  covModel = "BRR", predictor_Sec_complete = FALSE, nIter = 2500,
  burnIn = 500, thin = 5, progressBar = TRUE, digits = 4)
}
\arguments{
\item{data}{phenotipes matrix, with every trait in a independent column.}

\item{testingEnv}{environment to test.}

\item{ETA}{eta}

\item{covModel}{covariate model}

\item{predictor_Sec_complete}{FALSE by default}

\item{nIter}{number of iterations}

\item{burnIn}{number of burning}

\item{thin}{number of thinning}

\item{progressBar}{show the progress bar?}

\item{digits}{number of digits of accuracy in the results}
}
\value{
Something cool
}
\description{
Bayesian Multi-Output regression stacking for specific environment estimations
}
