% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMTME.R, R/BMTME_Package.R
\docType{package}
\name{BMTME}
\alias{BMTME}
\alias{BMTME-package}
\title{Bayes Multi-Trait Milti-Environment Model}
\usage{
BMTME(Y, X, Z1, Z2, nIter = 1000L, burnIn = 300L, thin = 2L,
  bs = ceiling(dim(Z1)[2]/6), parallelCores = 1, digits = 4,
  progressBar = TRUE, testingSet = NULL)
}
\arguments{
\item{Y}{no definition available.}

\item{X}{no definition available.}

\item{Z1}{no definition available.}

\item{Z2}{no definition available.}

\item{nIter}{no definition available.}

\item{burnIn}{no definition available.}

\item{thin}{no definition available.}

\item{bs}{no definition available.}

\item{parallelCores}{number of cores to use}

\item{digits}{number of digits of accuracy in the results}

\item{progressBar}{(Logical) Show the progress bar.}

\item{testingSet}{cv object}
}
\description{
The Bayesian Multi-Trait Multi-Environment models (BMTME) package was developed to implement...
}
