% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMTMERS.R
\name{BMTMERS}
\alias{BMTMERS}
\title{Bayesian Multi Trait Multi Environment Regressor Stacking}
\usage{
BMTMERS(Y = NULL, ETA = NULL, covModel = "BRR", nIter = 2500,
  burnIn = 500, thin = 5, progressBar = TRUE, testingSet = NULL,
  parallelCores = 1, digits = 4)
}
\arguments{
\item{Y}{phenotipes matrix, with every trait in a independent column.}

\item{ETA}{eta}

\item{covModel}{covariates model}

\item{nIter}{number of iterations}

\item{burnIn}{number of burning}

\item{thin}{number of thinning}

\item{progressBar}{The progress bar}

\item{testingSet}{cv object}

\item{parallelCores}{number of cores to use}

\item{digits}{number of digits of accuracy in the results}
}
\value{
Something cool
}
\description{
Bayesian Multi Trait Multi Environment Regressor Stacking
}
\examples{
\dontrun{
data('wheat')
CV.RandomPart(phenoWheat, NPartitions = 10, PTesting = 0.2, set_seed = 123)
}
}
