% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMTMERSEnv.R
\name{BMTMERS_Env}
\alias{BMTMERS_Env}
\title{Bayesian Multi Trait Multi Environment Regressor Stacking for specific environment estimations}
\usage{
BMTMERS_Env(data = NULL, testingEnv = "", ETA = NULL, nIter = 2500,
  burnIn = 500, thin = 5, progressBar = TRUE, digits = 4)
}
\arguments{
\item{testingEnv}{environment to test.}

\item{ETA}{eta}

\item{nIter}{number of iterations}

\item{burnIn}{number of burning}

\item{thin}{number of thinning}

\item{progressBar}{show the progress bar?}

\item{digits}{number of digits of accuracy in the results}

\item{Y}{phenotipes matrix, with every trait in a independent column.}
}
\value{
Something cool
}
\description{
Bayesian Multi Trait Multi Environment Regressor Stacking for specific environment estimations
}
\examples{
ETA=list(Env=list(X=Z.E,model="BRR"),Gen = list(X = Z.G, model = 'BRR'), EnvGen=list(X=Z.EG,model="BRR"))
testingSet <- BMTME::CV.RandomPart(pheno, NPartitions = 10, PTesting = 0.2, set_seed = 123)
}
