% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boxplot.BMECV}
\alias{boxplot.BMECV}
\title{boxplot.BMECV}
\usage{
\method{boxplot}{BMECV}(x, select = "Pearson", ordered = TRUE, ...)
}
\arguments{
\item{x}{\code{BMECV object} Objeto BMECV, resultado de ejecutar BME()}

\item{select}{\code{string} Pearson or MSEP}

\item{ordered}{\code{logic} TRUE or FALSE}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
