% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boxplot.BMORSCV}
\alias{boxplot.BMORSCV}
\title{boxplot.BMORSCV}
\usage{
\method{boxplot}{BMORSCV}(x, select = "Pearson", ordered = TRUE, ...)
}
\arguments{
\item{x}{\code{BMORSCV object} Objeto BMORSCV, resultado de ejecutar MTME()}

\item{select}{\code{string} Pearson or MAAPE}

\item{ordered}{\code{logic} TRUE or FALSE}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
