% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boxplot.BMTMECV}
\alias{boxplot.BMTMECV}
\title{boxplot.BMTMECV}
\usage{
\method{boxplot}{BMTMECV}(x, select = "Pearson", ordered = TRUE, ...)
}
\arguments{
\item{x}{\code{BMTMECV object} Objeto BMTMECV, resultado de ejecutar BMTME()}

\item{select}{\code{string} Pearson or MSEP}

\item{ordered}{\code{logic} TRUE or FALSE}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
