% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholesky.R
\name{cholesky}
\alias{cholesky}
\title{Cholesky}
\usage{
cholesky(a)
}
\arguments{
\item{a}{(numeric - matrix) an object to apply this method, it could be non positive-definite matrices.}
}
\description{
Compute the Cholesky factorization of a non-real symmetric positive-definite square matrix.
}
