% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomPartition.R
\name{CV.Stratified}
\alias{CV.Stratified}
\title{Cross-Validation with stratified samples}
\usage{
CV.Stratified(DataSet, NSamples = 10, nTesting = 10, replace = FALSE,
  set_seed = NULL)
}
\arguments{
\item{DataSet}{(\code{data.frame}) The object need contain three columns in the Tidy data format:
\code{$Line} is the Line or genotype identifier, and the name of this column could change.
\code{$Env} is the name of the evaluated environment (s).
\code{$Response} Variable response obtained for the row corresponding to line and environment.}

\item{NSamples}{(\code{integer}) Number of groups to the cross-validation.}

\item{nTesting}{(\code{integer}) A non-negative integer giving the number of items to choose.}

\item{replace}{(\code{logic})     Should sampling be with replacement?.}

\item{set_seed}{(\code{integer}) Number of seed for replicable research.}
}
\description{
Sample a number of observations to do a cross-validation, this method is similar like random partitions, but is stratified,
and could be used like bootstrapping sampling with the \code{replace} parameter as \code{TRUE}.
}
