% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{genoIranianToy}
\alias{genoIranianToy}
\title{Genomic values of Iranian toy dataset.}
\format{
30 x 30 matrix
}
\source{
International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}
\usage{
genoIranianToy
}
\description{
A data set based on a portion of the data used in the study of (Crossa et al., 2016).
}
\references{
Crossa, J., Jarquín, D., Franco, J., Pérez-Rodríguez, P., Burgueño, J., Saint-Pierre, C., Singh, S. (2016). Genomic Prediction of Gene Bank Wheat Landraces. G3: Genes|Genomes|Genetics, 6(7), 1819–1834. https://doi.org/10.1534/g3.116.029637
}
\keyword{Wheat}
\keyword{dataset}
