% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{genoMada}
\alias{genoMada}
\title{Genomic values of Mada dataset.}
\format{
50 x 50 matrix
}
\usage{
genoMada
}
\description{
A data set based on a portion of the data used in the study of (Ben Hassen et al., 2018).
}
\references{
Ben Hassen, M., Bartholomé, J., Valè, G., Cao, T.-V., & Ahmadi, N. (2018). Genomic Prediction Accounting for Genotype by Environment Interaction Offers an Effective Framework for Breeding Simultaneously for Adaptation to an Abiotic Stress and Performance Under Normal Cropping Conditions in Rice. G3: Genes|Genomes|Genetics, 8(July), 2319–2332. https://doi.org/10.1534/g3.118.200098
}
\keyword{Mada}
\keyword{dataset}
