% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{genoWheatToy}
\alias{genoWheatToy}
\title{Genomic values of Wheat toy dataset.}
\format{
30 x 30 data.frame
}
\source{
International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}
\usage{
genoWheatToy
}
\description{
A data set based on a portion of the data used in the study
from a collection of 599 historical CIMMYT wheat lines.
The wheat data set is from CIMMYT's Global Wheat Program.
}
\keyword{Wheat}
\keyword{dataset}
