% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{phenoJapa50}
\alias{phenoJapa50}
\title{Phenotypic values of Japa dataset.}
\format{
data.frame, 50 row per 4 columns.
}
\usage{
phenoJapa50
}
\description{
A data set based on a portion of the data used in the study of (Ben Hassen et al., 2018).
}
\references{
Ben Hassen, M., Bartholomé, J., Valè, G., Cao, T.-V., & Ahmadi, N. (2018). Genomic Prediction Accounting for Genotype by Environment Interaction Offers an Effective Framework for Breeding Simultaneously for Adaptation to an Abiotic Stress and Performance Under Normal Cropping Conditions in Rice. G3: Genes|Genomes|Genetics, 8(July), 2319–2332. https://doi.org/10.1534/g3.118.200098
}
\keyword{Japa}
\keyword{dataset}
