% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{phenoMaizeToy}
\alias{phenoMaizeToy}
\title{Phenotypic values of Maize dataset.}
\format{
data.frame, 90 row per 5 columns.
}
\usage{
phenoMaizeToy
}
\description{
A data set based on a portion of the data used in the study of (Montesinos-López et al., 2017).
}
\references{
Montesinos-López, O. A., Montesinos-López, A., Crossa, J., Montesinos-López, J. C., Luna-Vázquez, F. J., Salinas, J., … Buenrostro-Mariscal, R. (2017). A Variational Bayes Genomic-Enabled Prediction Model with Genotype × Environment Interaction. G3: Genes|Genomes|Genetics, 7(8), g3.117.041202. https://doi.org/10.1534/g3.117.041202
}
\keyword{Maize}
\keyword{dataset}
