% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{phenoWheatToy}
\alias{phenoWheatToy}
\title{Phenotypic values of Wheat toy dataset.}
\format{
data.frame, 90 row per 4 columns.
}
\source{
International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}
\usage{
phenoWheatToy
}
\description{
A data set based on a portion of the data used in the study
from a collection of 599 historical CIMMYT wheat lines.
The wheat data set is from CIMMYT's Global Wheat Program.
}
\keyword{Wheat}
\keyword{dataset}
