% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.BMORSCV}
\alias{plot.BMORSCV}
\title{Plot BMORSCV graph}
\usage{
\method{plot}{BMORSCV}(x, select = "Pearson", ...)
}
\arguments{
\item{x}{\code{BMORSCV object} BMORSCV object, result of use the MTME() function}

\item{select}{\code{character} By default ('Pearson'), plot the Pearson Correlations of the MTME Object, else ('MAAPE'), plot the MAAPE of the BMORSCV Object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot from BMORSCV object
}
