% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.BMTME}
\alias{plot.BMTME}
\title{plot.BMTME}
\usage{
\method{plot}{BMTME}(x, ...)
}
\arguments{
\item{x}{\code{BMTME object} Objeto BMTME, resultado de ejecutar BMTME}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
