% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.BMTMERSCV}
\alias{plot.BMTMERSCV}
\title{Plot BMTMERSCV graph}
\usage{
\method{plot}{BMTMERSCV}(x, select = "Pearson", ...)
}
\arguments{
\item{x}{\code{BMTMERSCV object} BMTMERSCV object, result of use the MTME() function}

\item{select}{\code{character} By default ('Pearson'), plot the Pearson Correlations of the MTME Object, else ('MSEP'), plot the MSEP of the BMTMERSCV Object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot from BMTMERSCV object
}
