% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.BMORS}
\alias{print.BMORS}
\title{Print BMORS information object}
\usage{
\method{print}{BMORS}(x, ...)
}
\arguments{
\item{x}{BMORS object}

\item{...}{more objects}
}
\value{
test
}
\description{
Print BMORS information object
}
