% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.BMORSCV}
\alias{print.BMORSCV}
\title{Print BMORSCV information object}
\usage{
\method{print}{BMORSCV}(x, ...)
}
\arguments{
\item{x}{object a}

\item{...}{more objects}
}
\value{
test
}
\description{
Print BMORSCV information object
}
