% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.BMTME}
\alias{print.BMTME}
\title{Print BMTME information object}
\usage{
\method{print}{BMTME}(x, ...)
}
\arguments{
\item{x}{object a}

\item{...}{more objects}
}
\value{
test
}
\description{
Print BMTME information object
}
