% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.BMTMECV}
\alias{print.BMTMECV}
\title{Print BMTMECV information object}
\usage{
\method{print}{BMTMECV}(x, ...)
}
\arguments{
\item{x}{object a}

\item{...}{more objects}
}
\value{
test
}
\description{
Print BMTMECV information object
}
