% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.BMTMERSCV}
\alias{print.BMTMERSCV}
\title{Print BMTMERSCV information object}
\usage{
\method{print}{BMTMERSCV}(x, ...)
}
\arguments{
\item{x}{object a}

\item{...}{more objects}
}
\value{
test
}
\description{
Print BMTMERSCV information object
}
