% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.BMTMERSENV}
\alias{print.BMTMERSENV}
\title{Print BMTMERSENV information object}
\usage{
\method{print}{BMTMERSENV}(x, ...)
}
\arguments{
\item{x}{object a}

\item{...}{more objects}
}
\value{
test
}
\description{
Print BMTMERSENV information object
}
