% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{residuals.BME}
\alias{residuals.BME}
\title{residuals.BME}
\usage{
\method{residuals}{BME}(object, digits = 4, ...)
}
\arguments{
\item{object}{\code{BME object} Objeto BME, resultado de ejecutar BME}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
