% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{residuals.BMTME}
\alias{residuals.BMTME}
\title{residuals.BMTME}
\usage{
\method{residuals}{BMTME}(object, digits = 4, ...)
}
\arguments{
\item{object}{\code{BMTME object} Objeto BMTME, resultado de ejecutar BMTME}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
