% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.BMECV}
\alias{summary.BMECV}
\title{Summary.BMECV}
\usage{
\method{summary}{BMECV}(object, information = "compact", digits = 4, ...)
}
\arguments{
\item{object}{\code{BMECV object} Objeto BMECV, resultado de ejecutar BME()}

\item{information}{compact, extended, complete}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
