% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.BMORSCV}
\alias{summary.BMORSCV}
\title{Summary.BMORSCV}
\usage{
\method{summary}{BMORSCV}(object, information = "compact", digits = 4,
  ...)
}
\arguments{
\item{object}{\code{BMORSCV object} Objeto BMORSCV, resultado de ejecutar MTME()}

\item{information}{compact, extended, complete}

\item{digits}{number of digits.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
