% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.BMORSENV}
\alias{summary.BMORSENV}
\title{Summary.BMORSENV}
\usage{
\method{summary}{BMORSENV}(object, digits = 4, ...)
}
\arguments{
\item{object}{\code{BMORSENV object} Objeto BMORSENV, resultado de ejecutar MTME()}

\item{digits}{number of digits.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
