% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.BMTME}
\alias{summary.BMTME}
\title{Summary.BMTME}
\usage{
\method{summary}{BMTME}(object, information = "compact", ...)
}
\arguments{
\item{object}{\code{BMTME object} Objeto BMTME, resultado de ejecutar BMTME}

\item{information}{compact, extended, complete}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
