% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.BMTMECV}
\alias{summary.BMTMECV}
\title{Summary.BMTMECV}
\usage{
\method{summary}{BMTMECV}(object, information = "compact", ...)
}
\arguments{
\item{object}{\code{BMTMECV object} Objeto BMTMECV, resultado de ejecutar BMTME()}

\item{information}{compact, extended, complete}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
