% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.BMTMERSCV}
\alias{summary.BMTMERSCV}
\title{Summary.BMTMERSCV}
\usage{
\method{summary}{BMTMERSCV}(object, information = "compact", digits = 4,
  ...)
}
\arguments{
\item{object}{\code{BMTMERSCV object} Objeto BMTMERSCV, resultado de ejecutar MTME()}

\item{information}{compact, extended, complete}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
