% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.BMTMERSENV}
\alias{summary.BMTMERSENV}
\title{Summary.BMTMERSENV}
\usage{
\method{summary}{BMTMERSENV}(object, digits = 4, ...)
}
\arguments{
\item{object}{\code{BMTMERSENV object} Objeto BMTMERSENV, resultado de ejecutar MTME()}

\item{...}{Further arguments passed to or from other methods.}

\item{information}{compact, extended, complete}
}
\description{
Solo es una prueba
}
