% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WheatJapa30}
\alias{WheatJapa30}
\title{WheatJapa30 Data}
\format{format, 300 row per 4 columns.}
\usage{
data(WheatJapa30)
}
\description{
The package includes a data set based on a portion of the data used in the study of
}
\author{
Luna-Vázquez, Francisco Javier
}
