% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WheatToy}
\alias{WheatToy}
\title{Phenotypic values and Genetic values of Wheat toy dataset.}
\format{format, 300 row per 4 columns.}
\usage{
data(WheatToy)
}
\description{
A data set based on a portion of the data used in the study of...
The variables are as follows:

* 'phenoWheatToy': Phenotypic values.
* 'genoWheatToy':  Geno values.
}
\author{
Luna-Vázquez, Francisco Javier
}
