% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{barplot.BMORSENV}
\alias{barplot.BMORSENV}
\title{barplot BMORSENV graph}
\usage{
\method{barplot}{BMORSENV}(height, select = "Pearson", ...)
}
\arguments{
\item{height}{an \code{BMORSENV object} for which the plot of model is meaningful.}

\item{select}{\code{character} By default ('Pearson'), plot the Pearson Correlations of the BMORSENV Object, else ('MAAPE'), plot the MAAPE of the BMORSENV Object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Creates a bar plot with vertical bars, showing the predictive capability of the model or the error rate.
}
