% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BME.R
\name{BME}
\alias{BME}
\title{Bayesian Multi-Environment Model (BME)}
\usage{
BME(Y, Z1, nIter = 1000L, burnIn = 300L, thin = 2L,
  bs = ceiling(dim(Z1)[2]/6), parallelCores = 1, digits = 4,
  progressBar = TRUE, testingSet = NULL)
}
\arguments{
\item{Y}{\code{(matrix)} Phenotypic respones where each column it's a different environment.}

\item{Z1}{\code{(matrix)} Matrix design for the genetic effects.}

\item{nIter}{\code{(integer)} Number of iterations to fit the model.}

\item{burnIn}{\code{(integer)} Number of items to burn at the beginning of the model.}

\item{thin}{\code{(integer)} Number of items to thin the model.}

\item{bs}{\code{(integer)} Number of groups.}

\item{parallelCores}{\code{(integer)} Number of cores to use.}

\item{digits}{\code{(integer)} Number of digits of accuracy in the results.}

\item{progressBar}{\code{(Logical)} Show the progress bar.}

\item{testingSet}{\code{(object or vector)} Crossvalidation object or vector with the positions to use like testing in a cross-validation test.}
}
\value{
If the testingSet is NULL, the function returns the predictions.

Else, if the testingSet is not NULL, the function returns the correlation of the predictions of the cross-validation test.
}
\description{
Bayesian Multi-Environment Model (BME)
}
