% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BME.R
\name{BME}
\alias{BME}
\title{BME}
\usage{
BME(Y, Z1, nIter = 1000L, burnIn = 300L, thin = 2L,
  bs = ceiling(dim(Z1)[2]/6), parallelCores = 1, digits = 4,
  progressBar = TRUE, testingSet = NULL)
}
\arguments{
\item{Y}{no definition available.}

\item{Z1}{no definition available.}

\item{nIter}{no definition available.}

\item{burnIn}{no definition available.}

\item{thin}{no definition available.}

\item{bs}{no definition available.}

\item{progressBar}{(Logical) Show the progress bar.}

\item{testingSet}{(numeric) Crossvalidation object or vector with the positions to use like testing in a cross-validation test.}
}
\value{
If the testingSet is NULL, the function returns the predictions.

Else, if the testingSet is not NULL, the function returns the correlation of the predictions of the cross-validation test.
}
\description{
BME
}
\examples{
Not example provided


}
