% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMORSEnv.R
\name{BMORS_Env}
\alias{BMORS_Env}
\title{Bayesian Multi-Output regression stacking for specific environment estimations}
\usage{
BMORS_Env(data = NULL, testingEnv = "", ETA = NULL,
  covModel = "BRR", predictor_Sec_complete = FALSE, nIter = 2500,
  burnIn = 500, thin = 5, progressBar = TRUE, digits = 4)
}
\arguments{
\item{data}{\code{(data.frame)} Phenotypic respones where each column it's a different trait and the first column are the name of the enviroment where it was evaluated.}

\item{testingEnv}{\code{(string)} Name of the Environment to test.}

\item{ETA}{\code{(matrix)} This is a two-level list used to specify the regression function (or linear predictor).}

\item{covModel}{\code{(string)} Name of the covariates model to implement (BRR, BayesA, BayesB, BayesC).}

\item{predictor_Sec_complete}{\code{(Logical)} FALSE by default.}

\item{nIter}{\code{(integer)} Number of iterations to fit the model.}

\item{burnIn}{\code{(integer)} Number of items to burn at the beginning of the model.}

\item{thin}{\code{(integer)} Number of items to thin the model.}

\item{progressBar}{\code{(Logical)} Show the progress bar.}

\item{digits}{\code{(integer)} Number of digits of accuracy in the results.}
}
\description{
Bayesian Multi-Output regression stacking for specific environment estimations
}
