% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMTME.R, R/BMTME_Package.R
\docType{package}
\name{BMTME}
\alias{BMTME}
\alias{BMTME-package}
\title{Bayes Multi-Trait Milti-Environment Model (BMTME)}
\usage{
BMTME(Y, X, Z1, Z2, nIter = 1000L, burnIn = 300L, thin = 2L,
  bs = ceiling(dim(Z1)[2]/6), parallelCores = 1, digits = 4,
  progressBar = TRUE, testingSet = NULL)
}
\arguments{
\item{Y}{\code{(matrix)} Phenotypic respones where each column it's a different trait.}

\item{X}{\code{(matrix)} Matrix design for the environment effects.}

\item{Z1}{\code{(matrix)} Matrix design for the genetic effects.}

\item{Z2}{\code{(matrix)} Matrix design for the genetic effects interaction with the environment effects.}

\item{nIter}{\code{(integer)} Number of iterations to fit the model.}

\item{burnIn}{\code{(integer)} Number of items to burn at the beginning of the model.}

\item{thin}{\code{(integer)} Number of items to thin the model.}

\item{bs}{\code{(integer)} Number of groups.}

\item{parallelCores}{\code{(integer)} Number of cores to use.}

\item{digits}{\code{(integer)} Number of digits of accuracy in the results.}

\item{progressBar}{\code{(Logical)} Show the progress bar.}

\item{testingSet}{\code{(object or vector)} Crossvalidation object or vector with the positions to use like testing in a cross-validation test.}
}
\description{
The Bayesian Multi-Trait Multi-Environment models (BMTME) package was developed to implement...
}
