% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMTMERS.R
\name{BMTMERS}
\alias{BMTMERS}
\title{Bayesian Multi Trait Multi Environment Regressor Stacking}
\usage{
BMTMERS(Y = NULL, ETA = NULL, nIter = 2500, burnIn = 500, thin = 5,
  progressBar = TRUE, CrossValidation = NULL, digits = 4)
}
\arguments{
\item{Y}{phenotipes matrix, with every trait in a independent column.}

\item{ETA}{eta}

\item{nIter}{number of iterations}

\item{burnIn}{number of burning}

\item{thin}{number of thinning}

\item{CrossValidation}{cv object}

\item{digits}{number of digits of accuracy in the results}
}
\value{
Something cool
}
\description{
Bayesian Multi Trait Multi Environment Regressor Stacking
}
\examples{
ETA=list(Env=list(X=Z.E,model="BRR"),Gen = list(X = Z.G, model = 'BRR'), EnvGen=list(X=Z.EG,model="BRR"))
CrossValidation <- BMTME::CV.RandomPart(pheno, NPartitions = 10, PTesting = 0.2, set_seed = 123)
}
