% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMTMERSEnv.R
\name{BMTMERS_Env}
\alias{BMTMERS_Env}
\title{Bayesian Multi Trait Multi Environment Regressor Stacking for specific environment estimations}
\usage{
BMTMERS_Env(data = NULL, testingEnv = "", ETA = NULL,
  covModel = "BRR", nIter = 2500, burnIn = 500, thin = 5,
  progressBar = TRUE, digits = 4)
}
\arguments{
\item{data}{phenotipes matrix, with every trait in a independent column.}

\item{testingEnv}{environment to test.}

\item{ETA}{eta}

\item{covModel}{covariate model}

\item{nIter}{number of iterations}

\item{burnIn}{number of burning}

\item{thin}{number of thinning}

\item{progressBar}{show the progress bar?}

\item{digits}{number of digits of accuracy in the results}
}
\value{
Something cool
}
\description{
Bayesian Multi Trait Multi Environment Regressor Stacking for specific environment estimations
}
