% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boxplot.BMTMECV}
\alias{boxplot.BMTMECV}
\title{boxplot.BMTMECV}
\usage{
\method{boxplot}{BMTMECV}(x, select = "Pearson", ordered = TRUE, ...)
}
\arguments{
\item{x}{an \code{BMTMECV object} for which the plot of model is meaningful.}

\item{select}{\code{character} By default ('Pearson'), plot the Pearson Correlations of the BMTMECV Object, else ('MAAPE'), plot the MAAPE of the BMTMECV Object.}

\item{ordered}{\code{logic} The graph should be sorted by the median? by default is \code{TRUE}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Produce box-and-whisker plot(s) of the given BMTMECV object.
}
