% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.BME}
\alias{plot.BME}
\title{plot.BME}
\usage{
\method{plot}{BME}(x, ...)
}
\arguments{
\item{x}{\code{BME object} Objeto BME, resultado de ejecutar BME}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
