% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.BME}
\alias{plot.BME}
\title{plot.BME}
\usage{
\method{plot}{BME}(x, trait = "", ...)
}
\arguments{
\item{x}{an \code{BME object} for which the plot of model is meaningful.}

\item{trait}{\code{string} Name of the trait to plot.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Simple scatter plot comparing the observed values against the predicted values.
}
