% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.BMORSCV}
\alias{plot.BMORSCV}
\title{Plot BMORSCV
Simple scatter plot comparing the observed values against the predicted values.}
\usage{
\method{plot}{BMORSCV}(x, select = "Pearson", ...)
}
\arguments{
\item{x}{an \code{BMORS object} for which the plot of model is meaningful.}

\item{select}{\code{character} By default ('Pearson'), plot the Pearson Correlations of the BMORSCV Object, else ('MAAPE'), plot the MAAPE of the BMORSCV Object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot BMORSCV
Simple scatter plot comparing the observed values against the predicted values.
}
