% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.BMTME}
\alias{plot.BMTME}
\title{plot.BMTME

Simple scatter plot comparing the observed values against the predicted values.}
\usage{
\method{plot}{BMTME}(x, trait = "", ...)
}
\arguments{
\item{x}{an \code{BMTME object} for which the plot of model is meaningful.}

\item{trait}{\code{string} Name of the trait to plot.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
plot.BMTME

Simple scatter plot comparing the observed values against the predicted values.
}
