% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.BME}
\alias{print.BME}
\title{Print BME information object}
\usage{
\method{print}{BME}(x, ...)
}
\arguments{
\item{x}{an BME object used to print.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print BME information object
}
