% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.BMECV}
\alias{print.BMECV}
\title{Print BMECV information object}
\usage{
\method{print}{BMECV}(x, ...)
}
\arguments{
\item{x}{an BMECV object used to print.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print BMECV information object
}
