% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.BMORSENV}
\alias{print.BMORSENV}
\title{Print BMORSENV information object}
\usage{
\method{print}{BMORSENV}(x, ...)
}
\arguments{
\item{x}{an BMORSENV object used to print.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
test
}
\description{
Print BMORSENV information object
}
