% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.BMTMECV}
\alias{print.BMTMECV}
\title{Print BMTMECV information object}
\usage{
\method{print}{BMTMECV}(x, ...)
}
\arguments{
\item{x}{an BMTMECV object used to print.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print BMTMECV information object
}
