% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{residuals.BME}
\alias{residuals.BME}
\title{residuals.BME}
\usage{
\method{residuals}{BME}(object, digits = 4, ...)
}
\arguments{
\item{object}{an \code{BME object} for which the extraction of model residuals is meaningful.}

\item{digits}{number of digits.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
extracts model residuals from BME objects returned by modeling function BME.
}
