% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{residuals.BMTME}
\alias{residuals.BMTME}
\title{residuals.BMTME}
\usage{
\method{residuals}{BMTME}(object, digits = 4, ...)
}
\arguments{
\item{object}{an \code{BMTME object} for which the extraction of model residuals is meaningful.}

\item{digits}{number of digits.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
extracts model residuals from BMTME objects returned by modeling function BMTME.
}
