% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.BMECV}
\alias{summary.BMECV}
\title{summary.BMECV}
\usage{
\method{summary}{BMECV}(object, information = "compact", digits = 4, ...)
}
\arguments{
\item{object}{\code{BMECV object} an BMECV object for which a summary is desired.}

\item{information}{The type of summary to obtain from the model (compact, extended, complete), by default is compact.}

\item{digits}{number of digits.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Produces a summary of the results of the fitted model adding
the predictive capabilities of the model, as well as the MAAPE error rate
and the respective standard errors.
}
