% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.BMORSENV}
\alias{summary.BMORSENV}
\title{summary.BMORSENV}
\usage{
\method{summary}{BMORSENV}(object, digits = 4, ...)
}
\arguments{
\item{object}{\code{BMORSENV object} an BMORSENV object for which a summary is desired.}

\item{digits}{number of digits.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Produces a summary of the results of the fitted model adding
the predictive capabilities of the model, as well as the MAAPE error rate
and the respective standard errors.
}
