% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WheatJapa30}
\alias{WheatJapa30}
\title{Wheat Japa 30 Data}
\format{
2 objects are loaded (phenoJapa30 and genoJapa30)
}
\usage{
data(WheatJapa30)
}
\description{
This data set is also based on the data used in the study of (Ben Hassen et al., 2018). The original dataset was composed of a sample of 167 lines evaluated for six traits each of them was evaluated in one environment. The total numbers of genome by sequencing (GBS) data were 32,066 SNPs and they were obtained with a heterozygosity rate <5% and minor allele frequency (MAF) > 5%. To load this dataset in the package we only use 30 lines, and we have identified this data set as WheatJapa30. For more details, see the study of (Ben Hassen et al., 2018). Also, in place of markers we provide with the GRM matrix.
}
\references{
Ben Hassen, M., Bartholomé, J., Valè, G., Cao, T.-V., & Ahmadi, N. (2018). Genomic Prediction Accounting for Genotype by Environment Interaction Offers an Effective Framework for Breeding Simultaneously for Adaptation to an Abiotic Stress and Performance Under Normal Cropping Conditions in Rice. G3: Genes|Genomes|Genetics, 8(July), 2319–2332. https://doi.org/10.1534/g3.118.200098
}
