% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WheatMadaToy}
\alias{WheatMadaToy}
\title{Wheat Mada Data}
\format{
2 objects are loaded (phenoMada and genoMada)
}
\usage{
data(WheatMadaToy)
}
\description{
This data set is based on the data used in the study of (Ben Hassen, Bartholomé, Valè, Cao, & Ahmadi, 2018). The original dataset was composed of a sample of 188 wheat lines evaluated for six traits. Each of them was evaluated in one environment. The total numbers of genome by sequencing (GBS) data were 32,066 single nucleotide polymorphisms (SNPs) that was obtained with a heterozygosity rate < 5% and minor allele frequency (MAF) > 5%. To load this dataset in the package we only use 30 lines, and we have identified this data set as WheatMadaToy. For more details, see the study of (Ben Hassen et al., 2018).
}
\references{
Ben Hassen, M., Bartholomé, J., Valè, G., Cao, T.-V., & Ahmadi, N. (2018). Genomic Prediction Accounting for Genotype by Environment Interaction Offers an Effective Framework for Breeding Simultaneously for Adaptation to an Abiotic Stress and Performance Under Normal Cropping Conditions in Rice. G3: Genes|Genomes|Genetics, 8(July), 2319–2332. https://doi.org/10.1534/g3.118.200098
}
